[#159696820] create log-in page
# [#159696820] Log-in page (UI Front-end) with the following fields.

First name
surname
Email
Password
Confirm password
submit button
sign up with other web services like
a. Facebook.
b. Google.
c. GitHub.
# Description of tasks to be completed
The sign-up/registration page for StackOverflow-lite with the registration form for a user to create an account.

# How should this be manually tested?
Clone or download this repository
https://github.com/jeffandeko/StackOverflow-lite
Then, navigate to the user-interface landing page and click the signup button on the main section or the one engraved in the header on click to access the sign-up page/registration form.
Any background context you want to provide?
HTML codes and CSS stylesheet

What are the relevant Pivotal Tracker Stories?
#159696230 https://www.pivotaltracker.com/story/show/159696230